requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator

# Vector database
chromadb

# AI and ML - Updated to compatible versions
sentence-transformers==2.2.2
transformers==4.35.2
torch==2.4.1
huggingface-hub==0.19.4

# PDF processing (optional)
PyPDF2==3.0.1

# HTTP requests
requests==2.31.0

# Environment and utilities
python-dotenv==1.0.0
python-multipart==0.0.6

# Async support
asyncio-mqtt==0.16.1

# Logging and utilities
colorama==0.4.6


